Move db folder to within app/ folder
