+++ app/db/knexClient.ts
@@ -1,6 +1,6 @@
 const knex = require('knex');
 
-const knexFile = require('../knexfile');
+const knexFile = require('../../knexfile');
 
 const environment = 'development';
