+++ app/db/seeds/dev/add_categories.ts
@@ -1,5 +1,5 @@
 import { Knex } from 'knex';
-import { categories } from '../../utils';
+import { categories } from '@/types';
 
 export async function seed(knex: Knex): Promise<void> {
   await knex('category').del();
